requirements.txt
fastapi
flake8
gunicorn
numpy
pyyaml
uvicorn
